Checkpoint
* Message handling should in theory work
* REST API communication works
* Working on consistent message passing and command handling

+++ msg.go
@@ -13,7 +13,7 @@ type Message struct {
 	HostID string
 	PeerID string
 	Type   string
-	Amount uint
+	Amount uint64
 }
 
 func serialize(msg *Message) []byte {
